chore: move 'newDisplayAmountFromNumber' out of primitives
