+++ src/domain/shared/primitives.ts
@@ -1,5 +1,3 @@
-import { getCurrencyMajorExponent } from "@domain/fiat"
-
 import { safeBigInt } from "./safe"
 
 // TODO: think how to differentiate physical from synthetic USD
@@ -104,27 +102,6 @@ export const displayAmountFromNumber = <T extends DisplayCurrency>({
   return {
     amount,
     currency,
-  }
-}
-
-export const newDisplayAmountFromNumber = <T extends DisplayCurrency>({
-  amount,
-  currency,
-}: {
-  amount: number
-  currency: T
-}): NewDisplayAmount<T> | ValidationError => {
-  const amountInMinor = safeBigInt(amount)
-  if (amountInMinor instanceof Error) return amountInMinor
-
-  const displayMajorExponent = getCurrencyMajorExponent(currency)
-
-  return {
-    amountInMinor,
-    currency,
-    displayInMajor: (Number(amountInMinor) / 10 ** displayMajorExponent).toFixed(
-      displayMajorExponent,
-    ),
   }
 }
